gif-converter-backend/requirements.txt
fastapi[all]
uvicorn
Pillow
requests
python-multipart
python-dotenv
numpy

